[no ci] docs: replace code fences with RST codeblocks.
If no own docstrings are provided for an overwritten method in a
subclass, the docstrings from the parent class are used.
As we inherit Keras objects, we sometimes get Keras docstrings,
which do not follow the conventions. Most importantly, they use
fenced code blocks instead of RST. This change replaces those
when building the documentation, making the docstrings somewhat
better to read.