[no ci] docs: more explicit make command names
There was some confusion arising from the fact that `make docs` did
not build the local docs. To avoid this confusion in the future,
the naming is made more explicit with the following changes:

make local -> make local-docs
make docs -> make production-docs
make docs-sequential -> make production-docs-sequential

The documentation and the GitHub actions are adapted as well.