Add class for custom transforms to adapter.
This commit reintroduces the features that were present in
`LambdaTransform`, but only allowing registered functions. While being
stricter, that allows for closer scaffolding and raising errors early
on, so that users cannot provide functions that will not be
(de)serializable later on.

As there are a few failure modes, the focus is on providing detailed
error messages to enable users to solve problems without external help.